adding collections directory
