+++ wlts/collections/class_system.py
@@ -5,25 +5,31 @@
 # Web Land Trajectory Service is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 #
-"""Python Class Module for WLTS."""
+"""WLTS Collection Abstract Class."""
+
+from wlts.datasources.ds_manager import datasource_manager
 
 class ClassificationSystemClass:
     """ClassificationSystemClass Class."""
 
     def __init__(self, **kwargs):
         """Creates a ClassificationSystemClass object."""
-        invalid_parameters = set(kwargs) - {"type", "name", "class_name", "id", "base", "code", "description"}
+        invalid_parameters = set(kwargs) - {"type", "name", "class_name", "value", "datasource_id"}
 
         if invalid_parameters:
             raise AttributeError('invalid parameter(s): {}'.format(invalid_parameters))
 
         self.type = kwargs['type']
+
         self.name = kwargs['name']
-        self.id =  kwargs['id']
+        self.value = kwargs['value']
         self.class_name = kwargs['class_name']
-        self.base = kwargs['base']
-        self.code = kwargs['code']
-        self.description = kwargs['description']
+
+        if self.type == 'Self':
+            self.datasource = None
+
+        else:
+            self.datasource = datasource_manager.get_datasource(kwargs['datasource_id'])
 
     def get_type(self):
         """Get ClassificationSystemClass type."""
@@ -33,22 +39,15 @@ class ClassificationSystemClass:
         """Get ClassificationSystemClass name."""
         return self.name
 
-    def get_description(self):
-        """Get ClassificationSystemClass description."""
-        return self.description
-
-    def get_id(self):
+    def get_value(self):
         """Get ClassificationSystemClass id."""
-        return self.id
+        return self.value
 
     def get_class_property_name(self):
         """Get ClassificationSystemClass property_name."""
         return self.class_name
 
-    def get_base(self):
-        """Get ClassificationSystemClass base : Workspace or schema."""
-        return self.base
+    def get_class_ds(self):
+        """Get DataSource of Classification System."""
+        return self.datasource
 
-    def get_code(self):
-        """Get ClassificationSystemClass workspace."""
-        return self.code
