--- wlts/json-config/wlts_config.json
@@ -1,72 +1,0 @@
-{
-  "datasources": {
-     "dbms_source": [
-      {
-        "type": "POSTGIS",
-        "id": "95b8acfa-5625-416e-a77a-b3e0f211553b",
-        "host": "localhost",
-        "port": "5432",
-        "user": "user",
-        "password": "password",
-        "database": "wlts"
-      }
-    ]
-  },
-  "feature_collection": [
-    {
-      "name": "sampledb",
-      "authority_name": "INPE",
-      "description": "Exemple Data",
-      "detail": "http://www.obt.inpe.br/",
-      "datasource_id": "95b8acfa-5625-416e-a77a-b3e0f211553b",
-      "dataset_type": "Table",
-      "classification_class": {
-        "type": "Feature",
-        "property_name": "wlts.sample_db_classes",
-        "property_id": "id",
-        "class_property_name": "nome",
-        "property_description": "descricao"
-      },
-      "temporal": {
-        "type": "STRING",
-        "resolution": {
-          "unit": "YEAR",
-          "value": "1"
-        },
-        "string_format": "%Y"
-      },
-      "period": {
-        "start_date": "2012",
-        "end_date": "2014"
-      },
-      "scala": "1:16.287.126",
-      "spatial_extent": {
-        "xmin":"-73.9905",
-        "ymin": "-34.7282",
-        "xmax": "-27.9904",
-        "ymax": "5.27184"
-      },
-      "feature_type": "wlts.sample_db",
-      "feature_id_property": "gid",
-      "geom_property": {
-        "property_name": "geom",
-        "srid": 4674,
-        "type": "Polygon"
-      },
-      "observations_properties": [
-        {
-          "class_property": "classe_2012",
-          "temporal_property": "2012"
-        },
-        {
-          "class_property": "classe_2013",
-          "temporal_property": "2013"
-        },
-        {
-          "class_property": "classe_2014",
-          "temporal_property": "2014"
-        }
-      ]
-    }
-  ]
-}
